Workaround: Do not set JSON_PREFIX on non-Darwin systems
Building libjson-rpc-cpp fails to find jsoncpp header files.  The reason
is that CMakeLists.txt sets JSON_PREFIX to jsoncpp, which is then passed
as JSONCPP_INCLUDE_PREFIX to the cmake project of libjson-rpc-cpp.  This
is the wrong prefix on distributions such as Arch Linux.

This commit removes the definition of JSON_PREFIX, leaving the cmake
project of libjson-rpc-cpp to figure it out on its own.